Changes to be committed: 	new file:   mcts.py 	new file:   pieces.py 	renamed:    state_machine_test.py -> pieces_test.py 	new file:   state.py 	modified:   state_machine.py 	new file:   state_test.py
